Reduced number of warnings
There are now only four sources of warnings:
Scene.cpp and Frontend.cpp (constructor order), GUI.cpp (line 119, no return from non-void function).
Fourth source is CEGUI which is difficult one.

+++ Riku/src/Frontend/Texture.cpp
@@ -20,7 +20,7 @@ void Texture::init(const std::string& type, const std::string& path, const std::
 	std::string path_tmp=path;
 	glGenTextures(1, &ID);
 		dir_tmp = std::filesystem::current_path().parent_path().string() + "/" + dir;
-		for(int i=0;i<path_tmp.size();i++)
+		for(int i=0;i<(int)path_tmp.size();i++)
 		{
 			if(path_tmp[i] == '\\')
 				path_tmp[i]='/';
