+++ Riku/src/GameLogic/Player/Player.cpp
@@ -9,14 +9,14 @@ Player::Player(int numberOfResources)
 
 void Player::acceptResources(int type, int quantity)
 {
-	if (type < 0 || type >= resources.size() || quantity <= 0)
+	if (type < 0 || type >= (int)resources.size() || quantity <= 0)
 		return;
 	resources[type] += quantity;
 }
 
 bool Player::useResources(int type, int quantity)
 {
-	if (type < 0 || type >= resources.size() || quantity <= 0 || resources[type] < quantity)
+	if (type < 0 || type >= (int)resources.size() || quantity <= 0 || resources[type] < quantity)
 		return false;
 	resources[type] -= quantity;
 	return true;
@@ -24,7 +24,7 @@ bool Player::useResources(int type, int quantity)
 
 int Player::getResourceQuantity(int type) const
 {
-	if (type < 0 || type >= resources.size() )
+	if (type < 0 || type >= (int)resources.size() )
 		return -1;
 	return resources[type];
 }
