+++ Riku/src/GameLogic/StateUpdate/PatchHandler/PlayerOnMovePatchHandler.cpp
@@ -2,7 +2,7 @@
 
 std::shared_ptr<IMove> PlayerOnMovePatchHandler::handlePatch(GameState& state, const Patch& patch) const
 {
-	if (patch.playerOnMove >= 0 && patch.playerOnMove < state.players.size())
+	if (patch.playerOnMove >= 0 && patch.playerOnMove < (int)state.players.size())
 		state.playerOnMove = patch.playerOnMove;
 	return nullptr;
 }
